requirements.txt
tensorflow-gpu>=1.11.0
numpy==1.15.4
opencv_python==3.4.3.18
Keras==2.2.4
Pillow==5.3.0
matplotlib>=3.0.0
